Refactor to better support multiple platforms and add cortexm to pytest suite
