+++ DeeployTest/test_generic_config.py
@@ -1,0 +1,68 @@
+# SPDX-FileCopyrightText: 2025 ETH Zurich and University of Bologna
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Test configuration for Generic platform."""
+
+KERNEL_TESTS = [
+    "Adder",
+    "MultIO",
+    "test1DConvolution",
+    "test2DConvolution",
+    "test1DDWConvolution",
+    "test2DDWConvolution",
+    "test1DPad",
+    "test2DPad",
+    "testGEMM",
+    "testMatMul",
+    "testMatMulAdd",
+    "testMaxPool",
+    "testRQConv",
+    "testRQMatMul",
+    "testReduceSum",
+    "testReduceMean",
+    "testSlice",
+    "testRequantizedDWConv",
+    "test2DRequantizedConv",
+    "iSoftmax",
+    "testFloatAdder",
+    "testFloatGEMM",
+    "testFloat2DConvolution",
+    "testFloat2DConvolutionBias",
+    "testFloat2DConvolutionZeroBias",
+    "testFloatLayerNorm",
+    "testFloatDiv",
+    "testFloat2DDWConvolution",
+    "testFloat2DDWConvolutionBias",
+    "testFloat2DDWConvolutionZeroBias",
+    "testFloatRelu",
+    "testFloatMaxPool",
+    "testFloatMatmul",
+    "testFloatReshapeWithSkipConnection",
+    "testFloatSoftmax",
+    "testFloatTranspose",
+    "testFloatMul",
+    "testFloatPowScalar",
+    "testFloatPowVector",
+    "testFloatSqrt",
+    "testFloatRMSNorm",
+    "Quant",
+    "Dequant",
+    "QuantizedLinear",
+]
+
+MODEL_TESTS = [
+    "simpleRegression",
+    "WaveFormer",
+    "simpleCNN",
+    "ICCT",
+    "ICCT_ITA",
+    "ICCT_8",
+    "ICCT_ITA_8",
+    "miniMobileNet",
+    "miniMobileNetv2",
+    "CCT/CCT_1_16_16_8",
+    "CCT/CCT_2_32_32_128_Opset20",
+    "testFloatDemoTinyViT",
+    "Autoencoder1D",
+]
